chore: moved scheduler config to core
- rename
